+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @scribe-sdk/cli
 
+## 0.1.0-alpha.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @scribe-sdk/styles@0.1.0-alpha.5
+  - @scribe-sdk/react@0.1.0-alpha.5
+  - @scribe-sdk/mdx@0.1.0-alpha.5
+
 ## 0.1.0-alpha.4
 
 ### Patch Changes